Update RSSdatabaseandtext.py
Added a function for actually doing error detection, came up with a few solutions for implementing error detection but have not started them yet really, the one that I have now is a temporary fix. recursion could be a issue with it but trying to implement a actual good solution soon.